Update glib binding > 2.66
Update based on https://github.com/gotk3/gotk3/pull/826

changes GetSource and GetTarget version restriction for <= 2.66
adding DupSource and DupTarget for version >=2.68
