+++ glib/gbinding.go
@@ -59,29 +59,11 @@ func (v *Binding) Unbind() {
 	C.g_binding_unbind(v.native())
 }
 
-// Retrieves the GObject instance used as the source of the binding
-func (v *Binding) GetSource() *Object {
-	obj := C.g_binding_get_source(v.native())
-	if obj == nil {
-		return nil
-	}
-	return wrapObject(unsafe.Pointer(obj))
-}
-
 // Retrieves the name of the property of “source” used as the source of
 // the binding.
 func (v *Binding) GetSourceProperty() string {
 	s := C.g_binding_get_source_property(v.native())
 	return C.GoString((*C.char)(s))
-}
-
-// Retrieves the GObject instance used as the target of the binding.
-func (v *Binding) GetTarget() *Object {
-	obj := C.g_binding_get_target(v.native())
-	if obj == nil {
-		return nil
-	}
-	return wrapObject(unsafe.Pointer(obj))
 }
 
 // Retrieves the name of the property of “target” used as the target of
